decouple url regex from matcher
