+++ src/Matcher/Matcher.php
@@ -4,11 +4,9 @@ declare(strict_types=1);
 
 namespace VStelmakh\UrlHighlight\Matcher;
 
-use VStelmakh\UrlHighlight\Url;
-
 /**
- * Finds URLs in a string via a URI regex, normalizes each match into a Url, and filters out
- * invalid hosts and trailing punctuation.
+ * Finds URLs in a string via {@see UrlRegex}, then discards the ones with an invalid host and trims trailing
+ * punctuation carried over from the surrounding text.
  *
  * @internal
  */
@@ -16,11 +14,13 @@ final readonly class Matcher
 {
     private PunctuationFilter $punctuationFilter;
     private HostValidator $hostValidator;
+    private UrlRegex $regex;
 
     public function __construct()
     {
         $this->punctuationFilter = new PunctuationFilter();
         $this->hostValidator = new HostValidator();
+        $this->regex = new UrlRegex();
     }
 
     /**
@@ -29,172 +29,17 @@ final readonly class Matcher
     public function match(string $text): array
     {
         $result = [];
-        $urlRegex = $this->regex();
-        /** @var $matches array<array{0: ?string, 1: int}> */
-        preg_match_all($urlRegex, $text, $matches, PREG_SET_ORDER + PREG_OFFSET_CAPTURE + PREG_UNMATCHED_AS_NULL);
-        foreach ($matches as $rawMatch) {
-            $url = $this->normalize($rawMatch);
-            if ($url !== null) {
-                $start = $rawMatch[0][1];
-                $result[] = new UrlMatch($start, $start + strlen($url->full), $url);
+
+        foreach ($this->regex->findAll($text) as $match) {
+            if (!$this->hostValidator->isValid($match->url)) {
+                continue;
             }
+
+            $url = $this->punctuationFilter->filter($match->url);
+
+            $result[] = $url === $match->url ? $match : new UrlMatch($match->start, $url);
         }
+
         return $result;
-    }
-
-    private function regex(): string
-    {
-        return implode("\n", [
-            '/',
-            $this->schemeRegex(),   // scheme, optional
-            $this->userinfoRegex(), // userinfo, optional
-            $this->hostRegex(),     // host, required
-            $this->portRegex(),     // port, optional
-            $this->pathRegex(),     // path, optional
-            $this->queryRegex(),    // query, optional
-            $this->fragmentRegex(), // fragment, optional
-            '/ixuJ',                // case-insensitive, extended, unicode, j-changed
-        ]);
-    }
-
-    private function schemeRegex(): string
-    {
-        return implode('', [
-            '(?|',                               // branch reset group
-                '(?<scheme>[a-z][a-z0-9+\-.]*)',     // start with letter, consists of: letter, number, "+", "-", "."
-                ':\/{2}',                            // followed by "://"
-                '|',                                 // or
-                '(?<scheme>mailto):',                // mailto, followed by ":"
-            ')?',                                // close group, optional
-        ]);
-    }
-
-    private function userinfoRegex(): string
-    {
-        return implode('', [
-            '(?:',                   // non-capturing group
-                '(?:',                   // non-capturing group
-                    '(?<=\/{2})',            // prefixed with "//" (has scheme)
-                    '|',                     // or
-                    '(?=[^',                 // lookahead, not starting with:
-                        '\p{Sm}',                // mathematical
-                        '\p{Sc}',                // currency
-                        '\p{Sk}',                // modifier symbol
-                        '\p{P}',                 // punctuation
-                    '])',                    // close lookahead
-                ')',                     // close group
-                '(?<userinfo>[',         // capturing group, only:
-                    '\p{L}\d\-\._~',         // unreserved
-                    '%',                     // percent encoded
-                    '!$&\'()*+,;=',          // sub-delims
-                    ':',                     // ":"
-                ']+)',                   // one or more, close group
-                '@',                     // suffixed with "@"
-            ')?',                    // close group, optional
-        ]);
-    }
-
-    private function hostRegex(): string
-    {
-        $label = implode('', [
-            '(?=[^\-])',                 // not start with: "-"
-            '(?:',                       // non-capturing group, consists of:
-                '[^',                        // not (exclude):
-                    '\p{Z}',                     // whitespace
-                    '\p{Sm}',                    // mathematical
-                    '\p{Sc}',                    // currency
-                    '\p{Sk}',                    // combining character (mark)
-                    '\p{C}',                     // control character (invisible)
-                    '\p{P}',                     // punctuation
-                ']',
-                '|',
-                '[',                         // except (include):
-                    '\-',                        // "-"
-                    '\x{200C}',                  // zero width non-joiner
-                    '\x{200D}',                  // zero width joiner
-                    '\x{00B7}',                  // middle dot
-                    '\x{0375}',                  // greek lower numeral sign
-                    '\x{05F3}',                  // hebrew punctuation geresh
-                    '\x{05F4}',                  // hebrew punctuation gershayim
-                    '\x{30FB}',                  // katakana middle dot
-                    '\x{0660}-\x{0669}',         // arabic-indic digits
-                    '\x{06F0}-\x{06F9}',         // extended arabic-indic digits
-                ']',
-            ')',                         // close group
-            '{1,63}',                    // length: 1-63 chars
-            '(?<=[^\-])',                // not end with: "-"
-        ]);
-
-        return implode('', [
-            '(?<host>',                       // capturing group
-                '(?<=\/{2}|@)',                   // prefixed with: "//" (has scheme) or "@" (email)
-                "(?:{$label}\.){0,}{$label}",     // no subdomain requirement
-                '|',                              // otherwise
-                "(?:{$label}\.){1,}{$label}",     // at least 1 subdomain
-            ')',                              // close group
-        ]);
-    }
-
-    private function portRegex(): string
-    {
-        return implode('', [
-            '(?:',              // non-capturing group
-                ':',                // prefixed with: ":"
-                '(?<port>\d+)',     // capturing group, at least 1 digit
-            ')?',               // close group, optional
-        ]);
-    }
-
-    private function pathRegex(): string
-    {
-        return implode('', [
-            '(?<path>',       // capturing group
-                '\/',             // prefixed with "/"
-                '[^\s\?\#]*',     // any chars except whitespace, "?", "#"
-            ')?',             // close group, optional
-        ]);
-    }
-
-    private function queryRegex(): string
-    {
-        return implode('', [
-            '(?<query>',    // capturing group
-                '\?',           // prefixed with "?"
-                '[^\s\#]*',     // any chars except whitespace, "#"
-            ')?',           // close group, optional
-        ]);
-    }
-
-    private function fragmentRegex(): string
-    {
-        return implode('', [
-            '(?<fragment>',  // capturing group
-                '\#',            // prefixed with "#"
-                '[^\s]*',        // any chars except whitespace
-            ')?',            // close group, optional
-        ]);
-    }
-
-    /**
-     * @param array<array{0: ?string, 1: int}> $rawMatch [0 => (string) value, 1 => (int) offset]
-     */
-    private function normalize(array $rawMatch): ?Url
-    {
-        $url = new Url(
-            full: $rawMatch[0][0] ?? '',
-            scheme: $rawMatch['scheme'][0],
-            userinfo: $rawMatch['userinfo'][0],
-            host: $rawMatch['host'][0] ?? '',
-            port: $rawMatch['port'][0] !== null ? (int) $rawMatch['port'][0] : null,
-            path: $rawMatch['path'][0],
-            query: $rawMatch['query'][0],
-            fragment: $rawMatch['fragment'][0],
-        );
-
-        if (!$this->hostValidator->isValid($url)) {
-            return null;
-        }
-
-        return $this->punctuationFilter->filter($url);
     }
 }
